chore: move host configuration to separate class
